% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSpecificLexiconEntry.R
\name{addSpecificLexiconEntry}
\alias{addSpecificLexiconEntry}
\title{Add a lexicon entry for an object of class fc}
\usage{
addSpecificLexiconEntry(fc, word, lemma)
}
\arguments{
\item{fc}{An object of class "fc".}

\item{word}{A word (character) to add in the specific lexicon defined in fc.}

\item{lemma}{An empty string if the word has to be removed, a synonym if the word has to be grouped with another word, or a group based on the hierarchical clustering.}
}
\value{
An object of class "fc".
}
\description{
The lexicon entry is saved in the specific lexicon tab of the lexicon xlsx file.
}
