% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDf.R
\name{cleanDf}
\alias{cleanDf}
\title{Clean a dataframe.}
\usage{
cleanDf(df, cols)
}
\arguments{
\item{df}{A dataframe.}

\item{cols}{A list of list, elements of the list being colName, dfName, type, clean, complete.}
}
\value{
The cleaned dataframe.
}
\description{
Check and clean a dataframe: column name and type check, removal of special characters, removal of duplicates, completion of missing values by NA.
}
\keyword{internal}
