% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToFile.R
\name{dfToFile}
\alias{dfToFile}
\title{Transform a dataframe into a HTML table.}
\usage{
dfToFile(df, format = "", fileName = "", caption = "", colSpec = list())
}
\arguments{
\item{df}{A dataframe.}

\item{format}{The format of the file to be saved: "html", "png", "pdf" or "jpeg".}

\item{fileName}{The name of the file to be saved.}

\item{caption}{Caption for the html table.}

\item{colSpec}{A list of list used to format the table columns, with following elements: index, name, width, background, bold, italic, color, border_left, border_right#'}
}
\value{
An object of class kable.
}
\description{
The dataframe can be styled and saved in an external file.
}
