% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceRate.R
\name{dominanceRate}
\alias{dominanceRate}
\title{Compute the dominance rate.}
\usage{
dominanceRate(
  dominances,
  significance = TRUE,
  boldComparison = FALSE,
  percentage = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{dominances}{Dataframe with columns: rep/product/subject/descriptor/time/score(0/1)/period}

\item{significance}{If TRUE, returns the significance}

\item{percentage}{If TRUE, returns the percentage, ifelse the absolute count}

\item{alpha}{Alpha risk for the binomial test.}
}
\value{
A list with dominanceRate, significanceLevel and parameters?
}
\description{
Compute the dominance rate.
}
