% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mr.sig.cell.R
\name{plot.mr.sig.cell}
\alias{plot.mr.sig.cell}
\title{Plot the table of mrCa test per cell}
\usage{
\method{plot}{mr.sig.cell}(
  res,
  alpha = 0.075,
  choice = "percent.derived.cont",
  col.greater = "green",
  col.lower = "orange",
  caption = "",
  format = "",
  fileName = "",
  ...
)
}
\arguments{
\item{res}{Result of mrCa test per cell}

\item{alpha}{TODO}

\item{choice}{TODO}

\item{col.greater}{TODO}

\item{col.lower}{TODO}

\item{caption}{TODO}

\item{format}{TODO}

\item{fileName}{TODO}

\item{...}{Further parameters}
}
\value{
a graph TODO
}
\description{
Return a HTMl object.
}
