% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.r
\name{plot_bar}
\alias{plot_bar}
\title{This function allows to draw spectra according to a group and potential smoothing}
\usage{
plot_bar(
  x,
  y,
  group = NULL,
  integrationTable = NULL,
  xlim = NULL,
  ylim = NULL,
  main = "Barplot",
  xlab = "Name",
  ylab = "Intensity",
  legend = TRUE,
  size = 1,
  ...
)
}
\arguments{
\item{x}{abscissa of the spectra}

\item{y}{ordinate of the spectra}

\item{group}{group for each spectra}

\item{integrationTable}{NULL}

\item{xlim}{vector of two numbers representing the limits of x axis.}

\item{ylim}{vector of two numbers representing the limits of y axis.}

\item{main}{title of the graph}

\item{xlab}{character indicating the label on the x axis}

\item{ylab}{character indicating the label on the y axis}

\item{legend}{if TRUE the legend is displayed}

\item{size}{size of the text}

\item{...}{Further parameters in plots such as xlim,xlab,ylim,ylab}
}
\description{
This function allows to draw spectra according to a group and potential smoothing
}
