% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileAnova.R
\name{profileAnova}
\alias{profileAnova}
\title{profileAnova}
\usage{
profileAnova(
  profileObject,
  model = "DefaultModel",
  randomEffects = c("Subject"),
  lsMeansAlpha = 0.1,
  lsMeansAdjustment = "Tukey",
  anovaCalculationMode = "Ols",
  show = c("Groups", "StatMANOVA", "F", "PValues", "RMSE"),
  fileName = NULL,
  varianceTest = "None",
  normalityTest = "None",
  orderByF = TRUE
)
}
\arguments{
\item{profileObject}{a profileObject obtained with \link{profileReadData}}

\item{model}{default to "DefaultModel". In this case, when ANOVA was calculated in profileReadData (useful for displaying statistics and groups on boxplot for example), the same model is kept for this function.}

\item{randomEffects}{default to c("Subject")}

\item{lsMeansAlpha}{default to 0.10. Alpha for the posthoc tests.}

\item{lsMeansAdjustment}{default to "Tukey"}

\item{anovaCalculationMode}{default to "Ols". Can also be "ML" for maximum likelihood}

\item{show}{default to c("Groups","StatMANOVA","F", "PValues","RMSE")}

\item{fileName}{default to NULL. In this case, the output file is called "ANOVA of scores". Otherwise, the output file is named as fileName.}

\item{varianceTest}{default to "None". Can also be "Bartlett" or "Fligner"}

\item{normalityTest}{default to "None". Can also be "ShapiroWilks"or "KolmogorovSmirnov"}

\item{orderByF}{default to TRUE. Indicates whether the attributes should be ordered by decreasing F or not.}
}
\value{
A list containing a table with ANOVA results and post hoc tests. The attributes are in rows and the products are in columns.
}
\description{
Allows to get the ANOVA results (and relative posthoc tests) for all descriptors
}
\examples{
#data(cheeses)
#profileAnova(cheeses)
}
