% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileBarPlot.R
\name{profileBarPlot}
\alias{profileBarPlot}
\title{profileBarPlot}
\usage{
profileBarPlot(
  profileObject,
  variable = profileObject$Attributes[[1]],
  itemsToPlot = c("GMean", "ErrorBars"),
  errorBars = "ConfInt",
  statsToPlot = c("Means", "Sd", "N"),
  groupId = "Both"
)
}
\arguments{
\item{profileObject}{profile object stemming from \link{profileReadData}}

\item{variable}{default to first attribute. Attribute to be considered.}

\item{itemsToPlot}{default to c("GMean","ErrorBars")}

\item{errorBars}{default to "ConfInt". Can also be "StdDev" or "StdErr"}

\item{statsToPlot}{default to c("Means","Sd","N"). A subset of these stats is also available}

\item{groupId}{default to "Both". Can also be "letters" or "colors".}
}
\description{
Returns barplots with ANOVA results and post-hoc tests (if selected in profileObject)
}
\examples{
data(cheeses)
profileBarPlot(profileObject=cheeses)
}
