% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileBoxPlot.R
\name{profileBoxPlot}
\alias{profileBoxPlot}
\title{profileBoxPlot}
\usage{
profileBoxPlot(
  profileObject,
  variable,
  overallBoxplot = TRUE,
  plotMeans = TRUE,
  plotStats = TRUE,
  groupId = "Both"
)
}
\arguments{
\item{profileObject}{profile object from \link{profileReadData}}

\item{variable}{descriptor to be plotted}

\item{overallBoxplot}{default to TRUE. Plot the overall variable (all products confounded)}

\item{plotMeans}{default to TRUE. Should the means be plotted with crosses ?}

\item{plotStats}{default to TRUE. Should the stats be displayed}

\item{groupId}{default to "both". Can also be "letters" or "colors".}
}
\description{
Returns the boxplot of descriptors per products with ANOVA and tests posthoc results
}
