% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileCVA.R
\name{profileCVA}
\alias{profileCVA}
\title{CVA}
\usage{
profileCVA(
  profileObject,
  axes = 1:2,
  test = "Hotelling-Lawley",
  option = "TwoWayANOVA",
  itemsToPlot = c("ProductEllipses", "ProductSegments"),
  representation = "distanceBiplot",
  nbAxes = "Auto",
  outputs = NULL,
  confInt = 0.9,
  ellipsesType = "barycentric",
  returnX = F,
  returnY = F,
  productColors = FALSE,
  fileName = NULL,
  ellipsesCalculation = "Chi",
  onlySignificantAttributesInMultivariateAnalysis = TRUE,
  alphaForSelectionOfSignificantAttributes = 0.05
)
}
\arguments{
\item{profileObject}{profile object from \link{profileReadData}}

\item{axes}{default to 1:2}

\item{test}{default to "Hotelling-Lawley".}

\item{option}{default to "TwoWayANOVA". Can also be "OneWayANOVA","TwoWayANOVA","MAM" or "MultiMAM".}

\item{itemsToPlot}{default to c("ProductEllipses","ProductSegments","Panellists")}

\item{representation}{default to "distanceBiplot"}

\item{nbAxes}{default to "Auto",}

\item{outputs}{default to NULL.}

\item{confInt}{default to 0.9.}

\item{ellipsesType}{default to "barycentric". can also be "individual".}

\item{returnX}{default to FALSE. If the X axis should be reverted, can be TRUE}

\item{returnY}{default to FALSE. If the Y axis should be reverted, can be TRUE}

\item{productColors}{default to FALSE}

\item{fileName}{default to NULL}

\item{ellipsesCalculation}{default to "Chi"}

\item{onlySignificantAttributesInMultivariateAnalysis}{default to TRUE. If FALSE, all attributes are considered.}

\item{alphaForSelectionOfSignificantAttributes}{default to 0.05. Significance threshold for the product effect for each attribute}
}
\value{
{
a list containing CVA results 
\itemize{
\item{IndivCoord }{coordinates of the products}
\item{VarCoord }{coordinates of the attributes}
\item{NbDimSig }{number of significant dimensions}
\item{HotellingTable }{table containing the results of T2 Hotelling tests for all product pairs}
\item{ConditioningOfW }{condition index of W. If ConditioningOfW >30, the matrix is ill-conditionned and CVA can be unstable}
\item{B }{covariance matrix of product effect}
\item{W}{covariance matrix of interaction}
\item{EigenVectors}{W-orthonormalized eigenvectors of W^(-1)B}
\item{EigenValues}{eigenvalues of W^(-1)B}
\item{Stats}{list containing the MANOVA statistics (statistics, approximated F and pvalue)}
\item{decomposition}{table with the decomposition of each score according to the different effects}
\item{CenteredProductSubjectTable}{centered product by subject table}
\item{nbAxes}{number of axes, useful if nbAxes="auto" was chosen}
\item{wDemi}{matrix of weigths used in the biplot}
\item{option}{returns the chosen model}
\item{representation}{if "biplot", the plot related to this CVA is a biplot. If else, it will be two graphs: a product map and an attribute map}
}

}
}
\description{
Return Canonical Variate Analysis
}
\examples{
data(cheeses)
profileCVA(cheeses)
}
