% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileDescriptiveStatistics.R
\name{profileDescriptiveStatistics}
\alias{profileDescriptiveStatistics}
\title{profileDescriptiveStatistics}
\usage{
profileDescriptiveStatistics(profileObject, by = "AttributeCode")
}
\arguments{
\item{profileObject}{profile object from \link{profileReadData}}

\item{by}{"AttributeCode" by default. Could also be "ProductCode".}
}
\description{
Returns some descriptive statistics (min, max, med...)
}
\examples{
data(cheeses)
profileDescriptiveStatistics(cheeses)
}
