% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileMAM.R
\name{profileMAM}
\alias{profileMAM}
\title{profileMAM}
\usage{
profileMAM(
  profileObject,
  output = "MAM.html",
  correctOnlyIfSignificant = FALSE,
  limitOfSignificance = 0.05,
  negativeCorrection = TRUE
)
}
\arguments{
\item{profileObject}{profileObject from \link{profileReadData}}

\item{output}{default to "MAM.html". Names of the output.}

\item{correctOnlyIfSignificant}{FALSE by default. If TRUE, scaling effect is corrected only when it is significant.}

\item{limitOfSignificance}{default to 0.05 limit of significance when correctOnlyIfSignificant is TRUE}

\item{negativeCorrection}{default to TRUE: the negative scaling are NOT taken into account in the sum squares. When FALSE, negative scaling are considered as scaling effect.}
}
\description{
Returns the results of Mixed Assessor Model
}
\examples{
data(cheeses)
profileMAM(cheeses)
}
