% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePCA.R
\name{profilePCA}
\alias{profilePCA}
\title{profilePCA}
\usage{
profilePCA(
  profileObject,
  axes = 1:2,
  option = "Covariance",
  representation = "DistanceBiplot",
  confInt = 0.9,
  repel = FALSE,
  nbAxes = "Auto",
  ellipsesType = "Barycentric",
  returnX = F,
  returnY = F,
  productColors = FALSE,
  ellipsesCalculation = "Chi",
  bootstrap = FALSE,
  nSamples = 100,
  twoDiffFiles = TRUE,
  itemsToPlot = c("ProductEllipses"),
  onlySignificantAttributesInMultivariateAnalysis = TRUE,
  alphaForSelectionOfSignificantAttributes = 0.05,
  sizeText = NULL,
  expandBiplot = 1,
  colorAttributeForBiplot = "black"
)
}
\arguments{
\item{profileObject}{profile object stemming from \link{profileReadData}}

\item{axes}{default to 1:2}

\item{option}{default to "Covariance". Could also be "Correlation"}

\item{representation}{default to "DistanceBiplot". Could also be "twoMaps"}

\item{confInt}{default to 0.9. Threshold for ellipses}

\item{repel}{if TRUE, ggrepel is computed. FALSE by default}

\item{nbAxes}{default to "Auto". Number of axes. When default, the number of axis is defined by...}

\item{ellipsesType}{default to "Barycentric". Can also be "Individual"}

\item{returnX}{default to FALSE. If the X axis should be reverted, can be TRUE}

\item{returnY}{default to FALSE. If the Y axis should be reverted, can be TRUE}

\item{productColors}{default to FALSE}

\item{ellipsesCalculation}{default to "Chi"}

\item{bootstrap}{default to FALSE: Warning, the ellipses are NOT obtained by bootstrap but with the covariance matrix of the individuals.}

\item{nSamples}{default to 100. If bootstrap is TRUE, number of bootstraps.}

\item{twoDiffFiles}{default to TRUE}

\item{itemsToPlot}{default to c("ProductEllipses","Panellists")}

\item{onlySignificantAttributesInMultivariateAnalysis}{default to TRUE}

\item{alphaForSelectionOfSignificantAttributes}{default to 0.05}

\item{sizeText}{number allowing to change the size of the text}

\item{expandBiplot}{default to 1.}

\item{colorAttributeForBiplot}{"black" by default}
}
\description{
Returns PCA.
}
\examples{
data(cheeses)
profilePCA(cheeses)

}
