% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileRadarPlot.R
\name{profileRadarPlot}
\alias{profileRadarPlot}
\title{profileRadarPlot}
\usage{
profileRadarPlot(
  profileObject,
  differentLines = "ProductCode",
  colors = NULL,
  minScale = 0,
  maxScale = 100
)
}
\arguments{
\item{profileObject}{profile object stemming from \link{profileReadData}}

\item{differentLines}{default to "ProductCode"}

\item{colors}{colors for the lines to be plotted}
}
\description{
Returns the radarplot for each descriptors
}
\examples{
data(cheeses)
profileRadarPlot(cheeses)
}
