% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileReadData.R
\name{profileReadData}
\alias{profileReadData}
\title{profileReadData}
\usage{
profileReadData(
  file = NULL,
  allData = NULL,
  outputsCsv = NULL,
  replaceNA = "crossmean",
  language = "en",
  sep = ";",
  transformSessionIntoReplicate = TRUE,
  percentageOfSubjectsByReplicate = 0,
  percentageOfReplicatesBySubject = 0,
  percentageOfProductBySubject = 0,
  productColors = NULL,
  additionalCalculations = T,
  model = NULL,
  randomEffects = "Subject",
  anovaCalculationMode = "Ols",
  lsMeansAdjustment = "Sidak",
  lsMeansAlpha = 0.05,
  onlySignificantAttributesInMultivariateAnalysis = T,
  alphaForSelectionOfSignificantAttributes = T
)
}
\arguments{
\item{file}{default to NULL. File to open whose colnames are "ProductCode" "SubjectCode","AttributeCode", "Replicate" and "Score" and potentially "Session"}

\item{allData}{default to NULL. Allows to use this function also for data already loaded in R. In this case, file should be NULL and allData should be a data.frame whose colnames are "ProductCode" "SubjectCode","AttributeCode", "Replicate" and "Score" (potentially also "Session")}

\item{replaceNA}{default to "crossmean"}

\item{sep}{default to ";". Colnames separator for the file}

\item{transformSessionIntoReplicate}{default to TRUE. Should a session be transformed to replicate ?}

\item{percentageOfSubjectsByReplicate}{default to 0. Option to deal with missing values.}

\item{percentageOfReplicatesBySubject}{default to 0. Option to deal with missing values.}

\item{percentageOfProductBySubject}{default to 0. Option to deal with missing values.}

\item{productColors}{default to NULL. Colors associated to each product (for radarplot)}

\item{additionalCalculations}{default to T. If TRUE, the anova calculations are calculated in the object and can be used in the graphical plot or further results.}

\item{model}{default to NULL. In this case, the model is "twoWayMultiplicative" for more than 1 replicate.}

\item{randomEffects}{default to "Subject".}

\item{anovaCalculationMode}{default to "Ols" (Ordinary Least Square). Could also be "ML" for maximum likelihood.}

\item{lsMeansAdjustment}{default to "Tukey". Could also be "sidak"}

\item{lsMeansAlpha}{default to 0.05. Limit of significance for product group constitutions}

\item{onlySignificantAttributesInMultivariateAnalysis}{default to TRUE. Should only the significant attributes be kept in multivariate analyses ?}

\item{alphaForSelectionOfSignificantAttributes}{default to T. Significance threshold for the attributes to be kept in the multivariate analyses.}
}
\description{
Read data from profile experiment
}
