% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXlsxOrCsv.R
\name{readXlsxOrCsv}
\alias{readXlsxOrCsv}
\title{Read an external file}
\usage{
readXlsxOrCsv(path, sep = ";")
}
\arguments{
\item{path}{Delimited text file, xls or xlsx file path}

\item{sep}{Decimal separator, if delimited text file}
}
\value{
Dataframe
}
\description{
Read an external file
}
