% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveReport.R
\name{saveReport}
\alias{saveReport}
\title{Save a list of outputs as an HTML report.}
\usage{
saveReport(x, fileName)
}
\arguments{
\item{x}{List of outputs}

\item{fileName}{Name of the HTML file}
}
\description{
Save a list of outputs as an HTML report.
}
